- Massive refactoring of the files. - Fixed my mistake previous fixes in this pull request.
