--- src/config.js
@@ -1,12 +1,0 @@
-let token, userId;
-
-const twitch = window.Twitch.ext;
-
-twitch.onContext((context) => {
-  console.log(context);
-});
-
-twitch.onAuthorized((auth) => {
-  token = auth.token;
-  userId = auth.userId;
-});
