+++ src/resources/fontawesome/webfonts/fa-v4compatibility.ttf
