+++ src/video_overlay.html
@@ -2,17 +2,16 @@
 <html>
 
     <head>
-        <title>Viewer Page</title>
+        <title>Ravenfall Overlay</title>
         <script src="https://extension-files.twitch.tv/helper/v1/twitch-ext.min.js"></script>
-        <script src="./jquery-3.3.1.min.js"></script>
-        <link rel="stylesheet" href="./default.css">
-        <link rel="stylesheet" href="./fontawesome/css/all.min.css">
+        <link rel="stylesheet" href="./resources/default.css">
+        <link rel="stylesheet" href="./resources/fontawesome/css/all.min.css">
     </head>
 
     <body style="font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;">
         <div id="app" class="full-height">
             <div class="extension-toggle" id="extension-toggle">
-                <img src="./imgs/ravenfall_logo_tiny.png" />
+                <img src="./resources/imgs/ravenfall_logo_tiny.png" />
             </div>
             <div class="extension hidden" id="extension">
                 <div class="extension-bg"></div>
@@ -21,13 +20,13 @@
                         <div class="toggler"></div>
                         <div class="value"><i class="fa fa-moon"></i></div>
                     </div>
-                    <img class="header" src="./imgs/ravenfall_logo2.png"></img>
+                    <img class="header" src="./resources/imgs/ravenfall_logo2.png"></img>
                     <div class="btn btn-close-panel"><i class="fa fa-close"></i></div>
                 </div>
 
                 <div class="views">
                     <div class="view" data-name="loading">
-                        <img src="./imgs/ravenfall_logo_tiny.png" class="no-connection" />
+                        <img src="./resources/imgs/ravenfall_logo_tiny.png" class="no-connection" />
                         <h2>Loading...</h2>
                         <p style="text-align: center;">Our ravens are checking if Ravenfall is ready to be played.</p>
                         <p class="show-if-id-unavailable" style="text-align: center;">Psst.. If nothing happens, try
@@ -36,14 +35,14 @@
 
                     <div class="view" data-name="error">
                         <div class="view sub-view" data-name="bad-server-connection">
-                            <img src="./imgs/sad-raven.png" class="no-connection" />
+                            <img src="./resources/imgs/sad-raven.png" class="no-connection" />
                             <h2>Unable to connect to the game server.</h2>
                             <p style="text-align: center;">We are retrying every couple of seconds. The server should be
                                 back online soon!</p>
                         </div>
 
                         <div class="view sub-view" data-name="game-not-running">
-                            <img src="./imgs/sad-raven.png" class="no-connection" />
+                            <img src="./resources/imgs/sad-raven.png" class="no-connection" />
                             <h2>Ravenfall is unavailable</h2>
                             <p style="text-align: center;">It looks like we don't have a connection to Ravenfall just
                                 yet,
@@ -51,7 +50,7 @@
                         </div>
 
                         <div class="view sub-view" data-name="anonymous-user">
-                            <img src="./imgs/sad-raven.png" class="no-connection" />
+                            <img src="./resources/imgs/sad-raven.png" class="no-connection" />
                             <h2>Twitch User Id is Required</h2>
                             <p style="text-align: center;">Unfortunately this extension is not able to verify who you
                                 are. Please make sure you grant the required permissions to share your user Id with this
@@ -72,19 +71,19 @@
                         </div>
 
                         <div class="view sub-view" data-name="initialize">
-                            <img src="./imgs/ravenfall_logo_tiny.png" class="no-connection" />
+                            <img src="./resources/imgs/ravenfall_logo_tiny.png" class="no-connection" />
                             <h2>Loading...</h2>
                             <p style="text-align: center;">Hold on to your Ravens..<br />You will be able to play soon!
                             </p>
                         </div>
                         <div class="view sub-view" data-name="create-account">
-                            <img src="./imgs/ravenfall_logo_tiny.png" class="no-connection" />
+                            <img src="./resources/imgs/ravenfall_logo_tiny.png" class="no-connection" />
                             <h2>You don't seem to have any characters.<br />First time playing?</h2>
                             <a class="btn btn-create-account">Click here or type !join in chat</a>
                         </div>
 
                         <div class="view sub-view" data-name="character-selection">
-                            <img src="./imgs/ravenfall_logo_tiny.png" class="no-connection" />
+                            <img src="./resources/imgs/ravenfall_logo_tiny.png" class="no-connection" />
                             <h2>Select character you want to play with</h2>
                             <div class="character-list"></div>
                         </div>
@@ -163,11 +162,11 @@
                                 </div>
 
                                 <div class="ferry-animation sailing">
-                                    <img src="./imgs/steamboat.gif" />
+                                    <img src="./resources/imgs/steamboat.gif" />
                                 </div>
 
                                 <div class="ferry-animation captain">
-                                    <img src="./imgs/disney-mickey.gif" />
+                                    <img src="./resources/imgs/disney-mickey.gif" />
                                 </div>
 
                                 <div class="dungeon-status">
@@ -414,10 +413,11 @@
             </div>
         </div>
 
-        <script src="./modules/debug.js"></script>
-        <script src="./modules/utilities.js"></script>
-        <script src="./modules/states.js"></script>
-        <script src="./modules/interactions.js"></script>
+        <script src="./resources/jquery-3.3.1.min.js"></script>
+        <script src="./resources/js/core.js"></script>
+        <script src="./resources/js/debug.js"></script>
+        <script src="./resources/js/interactions.js"></script>
+        <script src="./resources/js/states.js"></script>
         <script src="./extension.js" type="module"></script>
     </body>
